test_requirements.txt
MiniMock==1.2.8
Flask-Testing==0.4.2
mockredispy==2.9.0.10
nose==1.3.4
